CrossRef: Añadimos métodos en el DAO y el repositorio para gestionar la referencia cruzada entre ShoppingList e Item. Items por Categoría: Integramos la obtención de ítems por categoría en el repositorio de ítems

+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/domain/category/CategoryModel.kt
@@ -3,7 +3,7 @@ package ar.edu.unlam.mobile.scaffolding.domain.category
 import ar.edu.unlam.mobile.scaffolding.domain.item.ItemModel
 
 data class CategoryModel(
-    val id: Long,
+    val id: UInt,
     val name: String,
     val listItem: List<ItemModel> = emptyList(),
 )
